fix: Explicitly specify arguments in editor call
When a file needs to be opened, we create a command that opens the file
and call it. Previously, we were doing this with string concatenation,
but this introduces ambiguity when the filename contains whitespace,
e.g. is "my file.md" one file, or is it two?

This commit also changes some formatting.