feat: Add YAML serialization
Serializing a node using YAML is trivial. The trick is getting it back
into memory.

Note that this is slightly different from the official YAML
specification. I am adding attributes after the ":" separator, and am
using two spaces instead of four, and each node name is surrounded with
double quotes.